make bums a library, move examples to tests
