+++ memsafe-checker/Cargo.toml
@@ -1,12 +1,14 @@
 [package]
-name = "memsafe-checker"
+name = "bums"
 version = "0.1.0"
 edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+assert_cmd = "2.0.12"
 clap = { version = "4.4.6", features = ["derive"] }
 env_logger = "0.10.0"
 log = "0.4.20"
 rand = "0.8.5"
+predicates = "3.0.4"
